+++ src/libs/nav2_util/CMakeLists.txt
@@ -17,24 +17,39 @@ find_package(tf2 REQUIRED)
 
 include_directories(include)
 
-add_library(nav2_util STATIC
+add_library(costmap_lib STATIC
   src/costmap.cpp
+)
+
+add_library(map_lib STATIC
   src/map/map.c
   src/map/map_store.c
   src/map/map_range.c
   src/map/map_draw.c
   src/map/map_cspace.cpp
-
 )
 
-ament_target_dependencies(nav2_util
+add_library(pf_lib STATIC
+  src/pf/pf.c
+  src/pf/pf_kdtree.c
+  src/pf/pf_pdf.c
+  src/pf/pf_vector.c
+  src/pf/eig3.c
+  src/pf/pf_draw.c
+)
+
+ament_target_dependencies(costmap_lib
+  map_lib
+  pf_lib
   nav2_libs_msgs
   tf2_ros
   tf2
 )
 
 install(TARGETS
-  nav2_util
+  costmap_lib
+  map_lib
+  pf_lib
   ARCHIVE DESTINATION lib
   LIBRARY DESTINATION lib)
 
@@ -57,6 +72,6 @@ if(BUILD_TESTING)
 endif()
 
 ament_export_include_directories(include)
-ament_export_libraries(nav2_util)
+ament_export_libraries(costmap_lib map_lib pf_lib)
 
 ament_package()
