+++ src/libs/nav2_util/src/pf/pf_pdf.c
@@ -32,7 +32,7 @@
 //#include <gsl/gsl_rng.h>
 //#include <gsl/gsl_randist.h>
 
-#include "amcl/pf/pf_pdf.h"
+#include "nav2_util/pf/pf_pdf.h"
 
 // Random number generator seed value
 static unsigned int pf_pdf_seed;
