+++ src/libs/nav2_util/src/pf/pf_vector.c
@@ -30,8 +30,8 @@
 //#include <gsl/gsl_eigen.h>
 //#include <gsl/gsl_linalg.h>
 
-#include "amcl/pf/pf_vector.h"
-#include "amcl/pf/eig3.h"
+#include "nav2_util/pf/pf_vector.h"
+#include "nav2_util/pf/eig3.h"
 
 
 // Return a zero vector
