+++ src/nav2_localization/amcl/include/amcl/sensors/amcl_odom.h
@@ -30,7 +30,7 @@
 #define AMCL_ODOM_H
 
 #include "amcl_sensor.h"
-#include "../pf/pf_pdf.h"
+#include "nav2_util/pf/pf_pdf.h"
 
 namespace amcl
 {
